Extract `spec/concat.ts`
